backend/requirements.txt
flask_sqlalchemy
python-dotenv
psycopg2
flask
msal
pandas
flask-restx
sphinx
flask-cors